pyproject.toml
[tool.poetry]
name = "nba_stats_agent"
version = "0.1.0"
description = "Agentic NBA Stats Twitter Bot using Anthropic"
authors = ["Zak Hijaouy <zakhijaouy@gmail.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
anthropic = "^0.40.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 

[tool.poetry.scripts]
a = "src.main:main"

